Move all filter tests from test_spectrum_processor.py to test_all_filter_order.py
